requirements.txt
# Core ML and Data Science
tensorflow>=2.8.0
keras>=2.8.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Image Processing
opencv-python>=4.5.0
Pillow>=8.3.0
imageio>=2.9.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Web Interface
streamlit>=1.12.0
streamlit-option-menu>=0.3.0

# Progress and Utilities
tqdm>=4.62.0
click>=8.0.0
pyyaml>=6.0
python-dotenv>=0.19.0

# Medical Image Processing
pydicom>=2.2.0
nibabel>=3.2.0

# Model Interpretation
lime>=0.2.0
shap>=0.40.0

# Uncertainty Quantification
tensorflow-probability>=0.15.0

# Development and Testing
pytest>=6.2.0
pytest-cov>=3.0.0
black>=21.0.0
flake8>=4.0.0
mypy>=0.910

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0

# Deployment
gunicorn>=20.1.0
docker>=5.0.0

# Optional: GPU support
# tensorflow-gpu>=2.8.0  # Uncomment if using GPU